+++ src/runner.cuh
@@ -12,7 +12,7 @@ void cudaCheck(cudaError_t error, const char *file,
                int line); // CUDA error check
 void CudaDeviceInfo();    // print CUDA information
 
-void range_init_matrix(float *mat, int N);
+void range_init_matrix(float *mat, size_t N);
 void randomize_matrix(float *mat, int N);
 void zero_init_matrix(float *mat, int N);
 void copy_matrix(const float *src, float *dest, int N);
